[inductor] correctly handle resize for AOTInductor wrapper calls
When generating a wrapper call, we may have implicit resize applied to
the kernel's output. For example, for addmm(3d_tensor, 2d_tensor),
its output buffer is resized to a 2d tensor. This triggers a warning from
Aten's resize_output op:

    "UserWarning: An output with one or more elements was resized since it had...
    This behavior is deprecated, and in a future PyTorch release outputs will
    not be resized unless they have zero elements..."

More importantly, the output shape is not the same as we would expect, i.e.
2d tensor v.s. 3d tensor.

This PR fixed the issue by injecting resize_(0) before calling the relevant
kernel and resize_(expected_shape) after the kernel call.

We also fixed a minor typo in the PR.

ghstack-source-id: 7c065917b182dd68389fb267faf6e8e53f70a4c3
Pull Request resolved: https://github.com/pytorch/pytorch/pull/107848